# server/src/swagger/open-api.yaml
openapi: 3.0.0
info:
  title: Swagger UI
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: object
          description: The user id
        firstName:
          type: string
          description: The user first name
        lastName:
          type: string
          description: The user last name
        email:
          type: string
          format: email
          description: The user email
        password:
          type: string
          description: The user password-encrypted
        billingAddress:
          type: object
          properties:
            city:
              type: string
              description: The user city
            street:
              type: string
              description: The user street
        role:
          type: string
          description: The user role
      example:
        _id: 63ed5ea09980d3c74cfc9ae1
        firstName: first Name
        lastName: Last Name
        email: email@email.com
        password: tester
        billingAddress:
          type: object
          properties:
            city:
              type: string
              description: The user city
            street:
              type: string
              description: The user street
        role: admin
    UserRegistration:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
        - billingAddress
      properties:
        firstName:
          type: string
          description: The user first name
        lastName:
          type: string
          description: The user last name
        email:
          type: string
          format: email
          description: The user email
        password:
          type: string
          description: The user password
        billingAddress:
          type: object
          properties:
            city:
              type: string
              description: The user city
            street:
              type: string
              description: The user street
      example:
        firstName: first Name
        lastName: Last Name
        email: email@email.com
        password: tester
        billingAddress:
          city: test
          street: tested
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
          description: The user first name
        lastName:
          type: string
          description: The user last name
        email:
          type: string
          format: email
          description: The user email
        password:
          type: string
          description: The user password
        billingAddress:
          type: object
          properties:
            city:
              type: string
              description: The user city
            street:
              type: string
              description: The user street
      example:
        firstName: first Name
        lastName: Last Name
        email: email@email.com
        password: tester
        billingAddress:
          city: test
          street: tested
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: The user email
        password:
          type: string
          description: The user password-encrypted
      example:
        email: email@email.com
        password: tester
...
  responses:
    notFoundError:
      description: Not found error message
      content:
        text/plain:
          schema:
            type: string
            example: 'Not Found'
    internalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: Unexpected error
    badRequestError:
      description: Bad request error
      content:
        text/plain:
          schema:
            type: string
            example: Bad request error
    unauthorizedError:
      description: Unauthorized
      content:
        text/plain:
          schema:
            type: string
            example: Authorization header required
    forbiddenError:
      description: Forbidden
      content:
        text/plain:
          schema:
            type: string
            example: Forbidden
    user:
      description: user details
      content:
        application/json:
          schema:
            type: object
            $ref: '#/components/schemas/User'
...
    userLogin:
      description: User Login Details
      content:
        application/json:
          schema:
            type: object
            properties:
              user:
                $ref: '#/components/schemas/User'
              token:
                type: string
                description: Authorization Token
            example:
              user:
                _id: 63f268b4c17916bad7b0d0db
                firstName: first Name1
                lastName: Last Name1
                email: email1@email1.com
                password: $2a$10$K5E7230wX5VW9Mk2zFcBYODYpP1MflfEpGi4Bh0nO/31T/lC5UCD2
                role: admin
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImVtYWlsMUBlbWFpbDEuY29tIiwiaWF0IjoxNjc2OTMzODc4LCJleHAiOjE2NzY5MzQ0Nzh9.CyCdpILC9cjbR7W-s_XONHai8Zu-sFSitFlMglkcSsE
...
paths:
  /user:
    get:
      security:
        - bearerAuth: []
      summary: Returns specific user details
      tags:
        - Users
      parameters:
        - in: header
          name: customerEmail
          description: The user email
          schema:
            type: string
          required: true
      responses:
        '200':
          $ref: '#/components/responses/user'
        '400':
          $ref: '#/components/responses/badRequestError'
        '401':
          $ref: '#/components/responses/unauthorizedError'
        '403':
          $ref: '#/components/responses/forbiddenError'
        '404':
          $ref: '#/components/responses/notFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
  /user/register:
    post:
      security: []
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '200':
          $ref: '#/components/responses/userLogin'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      message:
                        type: string
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            message:
                              type: string
                            path:
                              type: string
                          required:
                            - message
                            - path
                  - type: object
                    properties:
                      status:
                        type: number
                      message:
                        type: string
        '500':
          $ref: '#/components/responses/internalServerError'
  /user/login:
    post:
      summary: Get user details and Authorizatoin Token
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      security: []
      responses:
        '200':
          $ref: '#/components/responses/userLogin'
        '400':
          description: Bad request error
          content:
            text/plain:
              schema:
                type: string
                example: incorrect username or password
        '500':
          $ref: '#/components/responses/internalServerError'
  /user/edit-profile:
    patch:
      security:
        - bearerAuth: []
      summary: edit user profile
      tags:
        - Users
      parameters:
        - in: header
          name: customerEmail
          description: The user email
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          $ref: '#/components/responses/userLogin'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      message:
                        type: string
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            message:
                              type: string
                            path:
                              type: string
                          required:
                            - message
                            - path
                  - type: object
                    properties:
                      status:
                        type: number
                      message:
                        type: string
        '500':
          $ref: '#/components/responses/internalServerError'
  /user/checkEmail:
    post:
      security: []
      summary: check if email exists in database
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: email to check
                  example: email@example.com
      responses:
        '200':
          description: email exists status
          content:
            application/json:
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/badRequestError'
        '401':
          $ref: '#/components/responses/unauthorizedError'
        '403':
          $ref: '#/components/responses/forbiddenError'
        '404':
          $ref: '#/components/responses/notFoundError'
        '500':
          $ref: '#/components/responses/internalServerError'
...